Button component with type predicates
