+++ src/components/Button.tsx
@@ -1,19 +1,21 @@
 import { type ComponentPropsWithoutRef } from "react";
 
-type ButtonProps = {
-	el: "button";
-} & ComponentPropsWithoutRef<"button">;
+type ButtonProps = ComponentPropsWithoutRef<"button"> & {
+	href?: never;
+};
 
-type AnchorProps = {
-	el: "anchor";
-} & ComponentPropsWithoutRef<"a">;
+type AnchorProps = ComponentPropsWithoutRef<"a"> & {
+	href?: string;
+};
 
-const Button = (props: ButtonProps | AnchorProps) => {
-	if (props.el === "anchor") {
+function isAnchorProps(props: ButtonProps | AnchorProps): props is AnchorProps {
+	return "href" in props;
+}
+
+export const Button = (props: ButtonProps | AnchorProps) => {
+	if (isAnchorProps(props)) {
 		return <a className="button" {...props}></a>;
 	}
 
 	return <button className="button" {...props}></button>;
 };
-
-export default Button;
