+++ src/components/Button_Dif_Elements.tsx
@@ -1,0 +1,17 @@
+import { type ComponentPropsWithoutRef } from "react";
+
+type ButtonProps = {
+	el: "button";
+} & ComponentPropsWithoutRef<"button">;
+
+type AnchorProps = {
+	el: "anchor";
+} & ComponentPropsWithoutRef<"a">;
+
+export const Button = (props: ButtonProps | AnchorProps) => {
+	if (props.el === "anchor") {
+		return <a className="button" {...props}></a>;
+	}
+
+	return <button className="button" {...props}></button>;
+};
